aggiunta popup errore e classi tab
