+++ Arduino/RxEmulator/RxEmulator.ino
@@ -65,43 +65,46 @@ void loop()
    Serial.println(td.ThrottlePedal_Pos);
   //td.ThrottlePedal_Pos = random(0,1000)/10.0;
   td.BrakePedal_Pos = 0;
-  td.SteeringWheel_Pos = random(0,1000)/10.0;
-  td.HVAcc_Voltage = random(0,1000)/10.0;
-  td.HVAcc_SoC = random(0,1000)/10.0;
-  td.HVAcc_Current = random(0,1000)/10.0;
-  td.HVAcc_Temp = random(0,1000)/10.0;
-  td.HVAcc_Cell_1_Temp = random(0,1000)/10.0;
-  td.HVAcc_Cell_2_Temp = random(0,1000)/10.0;
-  td.HVAcc_Cell_3_Temp = random(0,1000)/10.0;
-  td.HVAcc_Cell_4_Temp = random(0,1000)/10.0;
-  td.LVAcc_Voltage = random(0,240)/10.0;
-  td.LVAcc_SoC = random(0,1000)/10.0;
-  td.LVAcc_Current = random(0,200)/10.0;
-  td.inv_Temperature = random(0,1000)/10.0;
-  td.inv_HVVoltage = random(0,1000)/10.0;
-  td.inv_LVVoltage = random(0,240)/10.0;
-  td.MotorRR_ActCurrent = random(0,1000)/10.0;
-	//td.MotorRR_ActSpeed = random(0,200000)/10.0;
+  td.SteeringWheel_Pos = random(180,188);
+  td.HVAcc_Voltage = random(595, 605);
+  td.HVAcc_SoC = 94;
+  td.HVAcc_Current = random(110,120);
+  td.HVAcc_Temp = random(55,59);
+  td.HVAcc_Cell_1_Temp = random(55,59);
+  td.HVAcc_Cell_2_Temp = random(55,59);
+  td.HVAcc_Cell_3_Temp = random(55,59);
+  td.HVAcc_Cell_4_Temp = random(55,59);
+  td.LVAcc_Voltage = random(595, 605);
+  td.LVAcc_SoC = 87;
+  td.LVAcc_Current = random(110,120);
+  td.inv_Temperature =random(55,59);
+  td.inv_HVVoltage = random(595, 605);
+  td.inv_LVVoltage = random(23, 24);
+
+  td.MotorRR_ActCurrent = td.ThrottlePedal_Pos*1.2;
   td.MotorRR_ActSpeed = td.ThrottlePedal_Pos/10.0;
-	td.MotorRR_ActTorque = td.ThrottlePedal_Pos/10.0;
-	td.MotorRR_CmdSpeed = td.ThrottlePedal_Pos/10.0;
-	td.MotorRR_CmdTorque = td.ThrottlePedal_Pos/10.0;
-	td.MotorRL_ActCurrent = random(0,1000)/10.0;
+	td.MotorRR_ActTorque = td.ThrottlePedal_Pos;
+	td.MotorRR_CmdSpeed = td.ThrottlePedal_Pos*10;
+	td.MotorRR_CmdTorque = td.ThrottlePedal_Pos;
+  td.MotorRR_Frequency = td.ThrottlePedal_Pos*150;
+  td.MotorRR_Temperature = random(55,59);
+
+	td.MotorRL_ActCurrent = td.ThrottlePedal_Pos*1.2;
 	td.MotorRL_ActSpeed = td.ThrottlePedal_Pos/10.0;
-	td.MotorRL_ActTorque = td.ThrottlePedal_Pos/10.0;
-	td.MotorRL_CmdSpeed = td.ThrottlePedal_Pos/10.0;
-	td.MotorRL_CmdTorque = td.ThrottlePedal_Pos/10.0;
-	td.MotorRL_Frequency = random(0,1000)/10.0;
-	td.MotorRR_Frequency = random(0,1000)/10.0;
-	td.MotorRL_Temperature = random(0,1000)/10.0;
-	td.MotorRR_Temperature = random(0,1000)/10.0;
+	td.MotorRL_ActTorque = td.ThrottlePedal_Pos;
+	td.MotorRL_CmdSpeed = td.ThrottlePedal_Pos*10;
+	td.MotorRL_CmdTorque = td.ThrottlePedal_Pos;
+	td.MotorRL_Frequency = td.ThrottlePedal_Pos*150;
+	td.MotorRL_Temperature = random(55,59);
+
 	td.vehicle_linearSpeed = td.ThrottlePedal_Pos; 
-  td.saf_BSPD = random(0,100)>20;
-  td.saf_IMD = random(0,100)>20;
-  td.saf_LVMS = random(0,100)>20;
-	td.saf_AMS = random(0,100)>20;
-	td.saf_IS = random(0,100)>20;
-	td.saf_BOTS = random(0,100)>20;
+  //td.saf_BSPD = random(0,100)>20;
+  td.saf_BSPD = 100;
+  td.saf_IMD = 20;
+  td.saf_LVMS = 100;
+	td.saf_AMS = 20;
+	td.saf_IS = 20;
+	td.saf_BOTS = 100;
 	td.saf_SDBCockpit = random(0,100)>20;
 	td.saf_SDBLeft = random(0,100)>20;
 	td.saf_SDBRight = random(0,100)>20;
